improve filecontent to manage results more clearly
previously fileContentMatch returned [][]string, with the first
dimension being a line and the second the full match and any groups.
This has been replaced with a lineMatch type that isolates the full line
matched from any groups present. the groups that are present in the line
are used as criteria for evaluation as before.