add hasline source type and exactmatch evaluator
hasline is used to build true/false results for files identified that
match a specified name and contain a specified line.

the exactmatch evaluator is used to test if the criteria returned by a
source exactly matches the value. If so, the test is marked true for
that identifier.

These can be combined to test for files that just contain a given
expression, or files that do not have the expression.